requirements.txt
streamlit
transformers
langchain
PyPDF2
beautifulsoup4
requests
chromadb
openai
aiohttp
python-dotenv
tensorflow
torch 
torchvision 
torchaudio
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters